+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :feature:`-` ``PactGroup`` raises custom timeout exception from child if available
 * :release:`1.9.0 <6-8-2017>`
 * :bug:`- major` Handle exceptions thrown in a 'during' callback
 * :release:`1.8.0 <11-07-2017>`